requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-pandas==0.6.7
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.1.1
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
